+++ srcjs/shinyapp.js
@@ -833,8 +833,7 @@ var ShinyApp = function() {
   exports.progressHandlers = progressHandlers;
 
   // Returns a URL which can be queried to get values from inside the server
-  // function. This is enabled by calling `shiny::onTestSnapshot()` in the
-  // app's server function.
+  // function. This is enabled with `options(shiny.testmode=TRUE)`.
   this.getTestEndpointUrl = function() {
     return window.location.protocol + "//" + window.location.hostname +
       ":" + window.location.port + window.location.pathname +